refactor: modified file structures and style files
