+++ src/pages/AboutMe/components/Cards/Cards.style.ts
@@ -1,50 +1,7 @@
-import { onLarge, onSmall } from '@styles/mediaQuery'
+import { onSmall } from '@styles/mediaQuery'
 import styled from 'styled-components'
 
-export const HeroCardWarp = styled.div`
-  position: absolute;
-  top: 21.2rem;
-  min-width: 100rem;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  background-color: #212c4f;
-  padding: 5rem 8rem;
-  box-shadow: 0px 20px 40px rgba(0, 0, 0, 0.15);
-  border-radius: 14px;
-  box-sizing: border-box;
-`
-
-export const HeroCardInNavWrap = styled.div`
-  position: absolute;
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  z-index: 3;
-  top: 8.7rem;
-  width: 100%;
-  padding: 0 1.7rem;
-`
-
-export const IntroduceNameWarp = styled.div`
-  margin-bottom: 9rem;
-  font-size: 3.6rem;
-  line-height: 4.32rem;
-  font-weight: bold;
-  color: #fff;
-
-  ${onSmall} {
-    margin-bottom: 0;
-    font-size: 1.6rem;
-    line-height: 1.2;
-  }
-`
-export const DeveloperName = styled.span`
-  color: #91fccf;
-`
-export const IntroduceDevelopWarp = styled.div`
+export const IntroduceMeWrap = styled.div`
   color: white;
   font-size: 1.8rem;
   line-height: 1.5;
@@ -56,53 +13,15 @@ export const IntroduceDevelopWarp = styled.div`
     font-size: 1.6rem;
   }
 `
-export const DevelopTitle = styled.p`
+
+export const Title = styled.p`
   color: #4f83f9;
 `
-export const DevelopScr = styled.div`
+
+export const Description = styled.div`
   color: white;
   ${onSmall} {
     color: #205284;
-  }
-`
-export const IntroduceGithubWrap = styled.div`
-  display: flex;
-  flex-direction: column;
-  ${onSmall} {
-    position: relative;
-  }
-`
-export const AddImg = styled.img`
-  width: 22.3rem;
-  height: 22.3rem;
-  border-radius: 100%;
-  background-color: pink;
-  margin-bottom: 5rem;
-  ${onSmall} {
-    width: 9rem;
-    height: 9rem;
-    margin-bottom: 0rem;
-    position: absolute;
-    bottom: -1.5rem;
-    right: 18%;
-  }
-`
-export const BtnGithub = styled.button`
-  height: 6rem;
-  width: 22.8rem;
-  background: linear-gradient(93.96deg, #4b26ff 0%, #b01eff 100%);
-  box-shadow: 0px 10px 20px rgba(101, 41, 255, 0.15);
-  border-radius: 30px;
-  color: #fff;
-  font-size: 2.4rem;
-  font-weight: bold;
-  border: 0;
-  cursor: pointer;
-  ${onSmall} {
-    width: 6rem;
-    height: 2.5rem;
-    font-size: 1.2rem;
-    z-index: 4;
   }
 `
